fix(pddl): stop Pddl8Error::BoundExceeded.limit truncating past u8::MAX
`BoundExceeded { limit: u8, .. }` silently truncated any bound above
255: PDDL8_MAX_GROUND (4096) is the default grounding bound, and
`4096usize as u8 == 0`, so a domain that actually exceeded it reported
`limit: 0` in the refusal's Display/witness instead of the real 4096
-- misreporting the exact bound that was hit. A caller-supplied
`max_ground: Option<usize>` (GroundProblem::build, IndexedGroundProblem
::build) could truncate the same way for any value > 255, not just the
default constant.

Widen `limit` to `usize`, matching `got` and every constant that feeds
it, and drop the now-unnecessary `as u8` casts at all four call sites
(ground/mod.rs x3, ground/lazy.rs x1) plus the two sites that already
fit in u8 today but would silently break again if their constants ever
grew (parse.rs's PDDL8_MAX_ARITY/PDDL8_MAX_PARAMS/PDDL8_MAX_CONJUNCTS,
powl_bridge.rs's MAX_POWL_TAPE_STEPS). Updated the one existing test
that hardcoded the old u8 comparison (powl_bridge.rs), and the one
in-workspace consumer that redundantly re-cast the field back to usize
(nothing else in the workspace destructures `limit`).

Also updated crates/bcinr-pddl-lsp/src/bounds/mod.rs to drop its now-
redundant `as usize` cast for consistency -- that crate is excluded
from the workspace (external lsp-max-andon path dep unavailable in
this environment, see root Cargo.toml's comment), so this one file's
change is a same-pattern mechanical edit applied for consistency, not
verified by a build.

Added two regression tests (ground::bound_exceeded_tests): one with a
caller-supplied limit of 300 (> u8::MAX) and 301 ground actions, one
exercising the default PDDL8_MAX_GROUND path directly with 4097 ground
actions -- both assert the reported `limit` equals the real bound, not
`300usize as u8 == 44` / `4096usize as u8 == 0`.

Verified: cargo build --workspace (clean), cargo test -p bcinr-pddl
(87 unit + all integration suites pass, including the two new tests),
cargo clippy -p bcinr-pddl --all-targets --all-features -- -D warnings
(clean), cargo fmt -p bcinr-pddl (clean). bcinr-pddl-lsp's edit is
UNVERIFIED by build (crate excluded from the workspace).

+++ crates/bcinr-pddl/src/error.rs
@@ -25,9 +25,17 @@ pub enum Pddl8Error {
     /// PDDL text could not be parsed.
     ParseError(String),
     /// A structural bound was exceeded (arity, body atoms, variables, depth).
+    ///
+    /// `limit` is `usize`, matching `got` and every bound constant that
+    /// feeds it (`PDDL8_MAX_GROUND` = 4096, plus any caller-supplied
+    /// `max_ground: Option<usize>`) — it was previously `u8`, which silently
+    /// truncated any bound above 255 (`4096usize as u8 == 0`, so a domain
+    /// that actually exceeded `PDDL8_MAX_GROUND` reported `limit: 0` instead
+    /// of the real bound). A refusal's witness must report the bound that
+    /// was actually hit.
     BoundExceeded {
         what: &'static str,
-        limit: u8,
+        limit: usize,
         got: usize,
     },
     /// An action schema references an unknown predicate or type.
